Refactor messaging
