+++ server/api/message_apis.py
@@ -1,8 +1,9 @@
 from flask_restful import Resource, request, current_app
 from schematics.exceptions import DataError
+
 from server.models.dtos.message_dto import MessageDTO
 from server.services.authentication_service import token_auth, tm
-from server.services.message_service import MessageService, NotFound, MessageServiceError
+from server.services.messaging.message_service import MessageService, NotFound, MessageServiceError
 
 
 class ProjectsMessageAll(Resource):
